Add files via upload
This is a Work-out Wednesday Week 10 challenge in which i use a Violin chart custom visual in powerbi to visualize it a violin plot to visualize 
the average temperature in Denver, Colorado by month in 2020. 

One of the useful thing about the Violin plots is its useful for showing the 
distribution and density of data. An a typical Violin plots combines a kernel density plot with a box plot visualize a data.


Link to this dataset can be find at: https://bitbucket.org/dm-p/power-bi-visuals-violin-plot/downloads/
The challenge post can also be found at: http://www.workout-wednesday.com/pbi-2021-w10/

To know more about Custom Visual for the Violin plots: https://docs.microsoft.com/en-us/power-bi/developer/visuals/power-bi-custom-visuals