Create WOW2022_Week 12_Power BI Build a Slicer Panel
This challenge focused on a project management theme tracking. I tracked progress of project from data generated from the [MicrosoftPlanner.com](https://www.microsoft.com/en-ww/microsoft-365/business/task-management-software). Automating this planner to show tasks that are completed, in-progress and not-started  with Power BI was a really enlightening process for me